Changes related to Dto structure
