+++ backend/users/src/app/dto/appParticipants/viewAppParticipantsDto.ts
@@ -1,0 +1,11 @@
+import { Field, ObjectType } from '@nestjs/graphql';
+import { BaseAppParticipantsDto } from './baseAppParticipantsDto';
+
+@ObjectType()
+export class ViewAppParticipantsDto extends BaseAppParticipantsDto {
+  @Field()
+  id: number;
+
+  @Field()
+  applicationId: number;
+}
